# src/main/java/org/example/taskproject/dto/ResetPasswordDto.java
package org.example.taskproject.dto;



import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.Data;

@Data
public class ResetPasswordDto {
    @NotBlank(message = "Old password is required")
    private String oldPassword;

    @NotBlank(message = "New password is required")
    @Size(min = 6, message = "New password must be at least 6 characters")
    private String newPassword;

    @NotBlank(message = "Confirm new password is required")
    @Size(min = 6, message = "Confirm new password must be at least 6 characters")
    private String confirmNewPassword;
}